Move:add-leagueをpagesフォルダに移動 #27
